fix : couponQuantityIssued 발급 수량 Long타입으로 수정
feat : 유저 쿠폰 발급 기능 구현

+++ src/main/java/com/example/plusproject/domain/coupon/dto/request/CreateCouponRequestDto.java
@@ -28,6 +28,6 @@ public class CreateCouponRequestDto {
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
 	private final LocalDateTime couponEndDay;
 
-	private final int couponQuantityIssued;
+	private final Long couponQuantityIssued;
 
 }
